weather/pyproject.toml
[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project]
name = "weather"
version = "0.1.0"
description = "A toolkit for fetching real-time weather data and forecasts using OpenWeatherMap API."
requires-python = ">=3.10"
dependencies = [
    "arcade-tdk>=2.0.0,<3.0.0",
    "requests>=2.32.5",
    "python-dotenv>=1.0.0",
    "arcade-core>=2.3.0",
]
[[project.authors]]
name = "saeedbbm"


[project.optional-dependencies]
dev = [
    "arcade-ai[evals]>=2.1.4,<3.0.0",
    "arcade-serve>=2.0.0,<3.0.0",
    "pytest>=8.3.0,<8.4.0",
    "pytest-cov>=4.0.0,<4.1.0",
    "pytest-mock>=3.11.1,<3.12.0",
    "pytest-asyncio>=0.24.0,<0.25.0",
    "mypy>=1.5.1,<1.6.0",
    "pre-commit>=3.4.0,<3.5.0",
    "tox>=4.11.1,<4.12.0",
    "ruff>=0.7.4,<0.8.0",
    "types-requests>=2.31.0",
]
# Add LangChain as optional dependency for advanced demos
langchain = [
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-arcade>=0.1.0",
    "arcade-core>=2.0.0",
    "langgraph>=0.1.0",
]

# Tell Arcade.dev that this package is a toolkit
[project.entry-points.arcade_toolkits]
toolkit_name = "weather"

[tool.mypy]
files = [ "weather/**/*.py",]
python_version = "3.10"
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"
ignore_missing_imports = "True"

[tool.pytest.ini_options]
testpaths = [ "tests",]

[tool.coverage.report]
skip_empty = true

[tool.hatch.build.targets.wheel]
packages = [ "weather",]

[tool.uv]
dev-dependencies = [
    "weather[dev,langchain]",
]
